Start trying to define a language for the bot
The language would be based on what is currently present as chat
commands. The idea would be to be able to pass in scripts to run.

A simple way to allow arbitrary messages is to have a factory which
takes the raw command words and just returns a command if it can be
constructed from it. If no factory returns a command then the command is
unrecognized. The order of factory traversal would be the precedence of
the commands.

Eventually a proper parser would have to be written for this, if it were
taken to extremes. I want to have the language for this bot be high
level and declarative so I need to create an intermediate form first for
trying things out.