Provide access to things through their original names
The constants were to provide consistently named versions of them. To
load and save them to a persistent storage ideally only requires the
information they come with, which does not have these standardised
names.

At this point the constants will be used for clarity within the code.
All this may change in future.